+++ R/python.R
@@ -370,6 +370,7 @@ length.python.builtin.dict <- function(x) {
     
     names <- py_dict_get_keys_as_str(x)
     names <- names[substr(names, 1, 1) != '_']
+    Encoding(names) <- "UTF-8"
     types <- rep_len(0L, length(names))
     
   } else {
@@ -388,6 +389,7 @@ length.python.builtin.dict <- function(x) {
     name <- py_get_name(x)
     if (!is.null(name)) {
       submodules <- sort(py_list_submodules(name), decreasing = FALSE)
+      Encoding(submodules) <- "UTF-8"
       names <- c(names, submodules)
       types <- c(types, rep_len(5L, length(submodules)))
     }
@@ -788,7 +790,9 @@ py_list_attributes <- function(x) {
   ensure_python_initialized()
   if (py_is_module_proxy(x))
     py_resolve_module_proxy(x)
-  py_list_attributes_impl(x)
+  attrs <- py_list_attributes_impl(x)
+  Encoding(attrs) <- "UTF-8"
+  attrs
 }