+++ NEWS.md
@@ -9,6 +9,8 @@
 - Improved support for UTF-8 converstions (always use UTF-8 when converting from
   Python to R
 
+- Ignore private ("_" prefixed) attributes of dictionaries for .DollarNames
+
 
 # reticulate 1.1