+++ NEWS.md
@@ -12,7 +12,7 @@ Install the development version with: `install_github("rstudio/reticulate")`
 
 - Search in additional locations for Anaconda on Linux/Mac
 
-- Improved support for UTF-8 converstions (always use UTF-8 when converting from
+- Improved support for UTF-8 conversions (always use UTF-8 when converting from
   Python to R
 
 - Ignore private ("_" prefixed) attributes of dictionaries for .DollarNames